Refactored code to make testing fun again. Said no one ever.
