+++ src/component/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../style/App.css';
-import ComicList from '../container/ComicList';
+import ItemList from '../container/ItemList';
 import ComicDetail from '../component/ComicDetail';
 import Navbar from '../container/Navbar';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
@@ -11,7 +11,7 @@ function App() {
     <Router>
       <div className="App">
         <Navbar />
-        <Route exact path="/"  component={ComicList} />
+        <Route exact path="/"  component={ItemList} />
         <Route exact path="/comic/:key" component={ComicDetail}/>
         <Footer />
      </div>
